+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Beanstalkd lib
 
+## Unreleased
+
+* **[ADDED]** `Zlikavac32\BeanstalkdLib\Protocol\TraceableProtocol` and `Zlikavac32\BeanstalkdLib\Adapter\PHPUnit\Constraint\OrderedTracesExistInProtocol` to help with testing
+
 ## 0.1.1 (2019-04-26)
 
 * **[CHANGED]** Bump `zlikavac32/alarm-scheduler` to `^0.2.1`